requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-rest-auth==1.0.1
Django==3.0.5
django-allauth==0.41.0
django-debug-toolbar==2.2
django-filter==2.2.0
django-rest-swagger==2.2.0
django-versatileimagefield==2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
entrypoints==0.3
flake8==3.7.9
idna==2.9
inflection==0.4.0
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
Pillow==7.1.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-magic==0.4.15
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wrapt==1.11.2
